fix: decouple inbound prompt context from runner parameters
