+++ src/auto-reply/reply/queue/types.ts
@@ -4,10 +4,8 @@ import type { QueueMode } from "../../../../packages/gateway-protocol/src/schema
 import type { AutoFallbackPrimaryProbe } from "../../../agents/agent-scope.js";
 import type { ExecToolDefaults } from "../../../agents/bash-tools.js";
 import type { CliSessionBindingFacts } from "../../../agents/cli-runner/types.js";
-import type {
-  CurrentInboundPromptContext,
-  RunEmbeddedAgentParams,
-} from "../../../agents/embedded-agent-runner/run/params.js";
+import type { RunEmbeddedAgentParams } from "../../../agents/embedded-agent-runner/run/params.js";
+import type { CurrentInboundPromptContext } from "../../../agents/internal-runtime-context.js";
 import type { ModelFallbackRouteResolution } from "../../../agents/model-fallback.types.js";
 import type { ScheduledToolPolicyContext } from "../../../agents/scheduled-tool-policy.js";
 import type { TrustedSubagentCompletionHandoff } from "../../../agents/subagents/announce/subagent-announce-handoff.js";
