+++ src/auto-reply/reply/reply-run-registry.contracts.ts
@@ -1,4 +1,4 @@
-import type { CurrentInboundPromptContext } from "../../agents/embedded-agent-runner/run/params.js";
+import type { CurrentInboundPromptContext } from "../../agents/internal-runtime-context.js";
 import type { ScheduledToolPolicyContext } from "../../agents/scheduled-tool-policy.js";
 import type { TrustedSubagentCompletionHandoff } from "../../agents/subagents/announce/subagent-announce-handoff.js";
 import type { ChatType } from "../../channels/chat-type.js";
